put SAVE_TIMING_DIR_PROJECTS in env_run.xml
With the simplified implementation of is_save_timing_dir_project,
the dependence on config_machines.xml can be eliminated after
the case is created by putting SAVE_TIMING_DIR_PROJECTS in, and
reading it from, env_run.xml . This is more consistent with
how other parameters of this sort are handled.

[BFB]
