+++ cime/scripts/lib/CIME/XML/machines.py
@@ -260,28 +260,6 @@ class Machines(GenericXML):
         return mpilib == "mpi-serial" or \
             self.get_field_from_list("MPILIBS", reqval=mpilib, attributes=attributes) is not None
 
-    def is_save_timing_dir_project(self,project):
-        """
-        Check whether the project is permitted to archive performance data in the location specified for the current machine
-
-        >>> machobj = Machines(machine="edison")
-        >>> machobj.is_save_timing_dir_project("acme")
-        True
-        >>> machobj.is_save_timing_dir_project("ccsm1")
-        False
-        """
-        save_timing_dir_projects = self.get_value("SAVE_TIMING_DIR_PROJECTS")
-        if not save_timing_dir_projects:
-            return False
-        else:
-            save_timing_dir_projects = save_timing_dir_projects.split(",")
-            for save_timing_dir_project in save_timing_dir_projects:
-                regex = re.compile(save_timing_dir_project)
-                if regex.match(project):
-                    return True
-
-            return False
-
     def has_batch_system(self):
         """
         Return if this machine has a batch system
