+++ cime/scripts/lib/CIME/provenance.py
@@ -5,7 +5,6 @@ Library for saving build/run provenance.
 """
 
 from CIME.XML.standard_module_setup import *
-from CIME.XML.machines      import Machines
 from CIME.utils import touch, gzip_existing_file, SharedArea, copy_umask
 
 import tarfile, getpass, signal, glob, shutil
@@ -95,13 +94,11 @@ def save_build_provenance(case, lid=None):
             _save_build_provenance_cesm(case, lid)
 
 def _save_prerun_timing_acme(case, lid):
-    mach = case.get_value("MACH")
-    machobj = Machines(machine=mach)
     project = case.get_value("PROJECT", subgroup="case.run")
-    if not machobj.is_save_timing_dir_project(project):
+    if not case.is_save_timing_dir_project(project):
         return
 
-    timing_dir = machobj.get_value("SAVE_TIMING_DIR")
+    timing_dir = case.get_value("SAVE_TIMING_DIR")
     if timing_dir is None or not os.path.isdir(timing_dir):
         logger.warning("SAVE_TIMING_DIR {} is not valid. E3SM requires a valid SAVE_TIMING_DIR to archive timing data.".format(timing_dir))
         return
@@ -123,6 +120,7 @@ def _save_prerun_timing_acme(case, lid):
         logger.warning("{} cannot be created. Skipping archive of timing data and associated provenance.".format(full_timing_dir))
         return
 
+    mach = case.get_value("MACH")
     compiler = case.get_value("COMPILER")
 
     # For some batch machines save queue info
@@ -286,16 +284,15 @@ def _save_postrun_timing_acme(case, lid):
     timing_saved_file = "timing.%s.saved" % lid
     touch(os.path.join(caseroot, "timing", timing_saved_file))
 
-    mach = case.get_value("MACH")
-    machobj = Machines(machine=mach)
     project = case.get_value("PROJECT", subgroup="case.run")
-    if not machobj.is_save_timing_dir_project(project):
+    if not case.is_save_timing_dir_project(project):
         return
 
-    timing_dir = machobj.get_value("SAVE_TIMING_DIR")
+    timing_dir = case.get_value("SAVE_TIMING_DIR")
     if timing_dir is None or not os.path.isdir(timing_dir):
         return
 
+    mach = case.get_value("MACH")
     base_case = case.get_value("CASE")
     full_timing_dir = os.path.join(timing_dir, "performance_archive", getpass.getuser(), base_case, lid)
